secure --cyclic
There are signs that even when dutyEnabled is unset, making
sleep/record seconds be 0/0 is a bad idea.  The official tools
default to 0/1 or 5/55.

Unclear if this change was needed: I haven't tried disabling it,
and my problems at the time may have been caused by the real
config encoding error having to do with 4/5 time periods.